@@ -28,7 +28,7 @@ Dummy::Application.configure do
   # The :test delivery method accumulates sent emails in the
   # ActionMailer::Base.deliveries array.
   config.action_mailer.delivery_method = :test
-  ActionMailer::Base.default :from => "no-reply@foo.com>"
+  ActionMailer::Base.default from: "no-reply@foo.com>"
 
   # Raise exception on mass assignment protection for Active Record models
   config.active_record.mass_assignment_sanitizer = :strict
